split processTracking into header and source
